Finishing testing on small examples the effect of basic heuristic to break up giant components.

+++ core/inc/cactus_core.h
@@ -37,7 +37,7 @@ typedef struct _CactusCoreInputParameters {
     bool singleCopyIngroup;
     bool singleCopyOutgroup;
 
-    double maxAdjacencyComponentSizeRatio;
+    float maxAdjacencyComponentSizeRatio;
 } CactusCoreInputParameters;
 
 int32_t cactusCorePipeline(Flower *flower, CactusCoreInputParameters *cCIP,
